Remove cached signature validations for an inode when it is freed.
Keeping it around is invalid, as the next time the inode is read, the
inode * will be at a different memory location, causing a different hash
value for the inode.

If that were *not* the case, that would be even worse, as then we would
have persistant cached validations.  As pointed out by Marco Slaviero,
removing a cd/floppy and reinserting a new one would be a potential
way to avoid signature validation.