digsig: init module using module_init()
Initialize module with module_init() rather than security_initcall().

Although the module should be initialized with security_initcall() like
the rest of the kernel security modules, it seems we cannot call
digsig_init_sysfs() so early on in the kernel initialization without
causing a panic due to a null pointer dereference.

Eventually, the module should use securityfs for communication with
user space.

Signed-off-by: Rupert Muchembled <rupert@rmuch.net>