digsig: check for usb devices, blacklisted filesystems
Change is_unprotected_file() in digsig.c to check if a file is stored
on a USB mass storage device. Executing a binary on an untrusted USB
mass storage device can present a security risk, because the file can
be written while it is being read by malicious firmware on the USB
storage controller.

This patch also adds an array of blacklisted filesystems, rather than
a simple check for just NFS, since there are network filesystems other
than NFS which present a similar security risk.

Signed-off-by: Rupert Muchembled <rupert@rmuch.net>